requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.8.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.11.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.19.0
pytz==2024.1
tzdata==2025.1